+++ R/MazamaCoreUtils.R
@@ -1,6 +1,6 @@
 #' @docType package
 #' @name MazamaCoreUtils
-#' @title Mazama Science utility functions for building operational systems.
+#' @title Utility Functions for Production R Code
 #' @description The MazamaCoreUtils package was created by MazamaScience to
 #' regularize our work building R-based web services.
 #'