+++ R/parseDatetime.R
@@ -25,12 +25,12 @@
 #' If \code{datetime} is a \code{POSIXct} it will be returned unmodified, and
 #' formats not recognized will be returned as \code{NA}.
 #'
-#' @param datetime vector of character or integer datetimes in Ymd[HMS] format
+#' @param datetime Vector of character or integer datetimes in Ymd[HMS] format
 #'   (or POSIXct).
-#' @param timezone Olson timezone at the location of interest.
+#' @param timezone Olson timezone used to interpret dates (required).
 #' @param expectAll Logical value determining if the function should fail if
 #'   any elements fail to parse (default \code{FALSE}).
-#' @param julian Logical value determining whether \code{datetime} should be
+#' @param isJulian Logical value determining whether \code{datetime} should be
 #' interpreted as a Julian date with day of year as a decimal number.
 #' @param quiet Logical value passed on to \code{lubridate::parse_date_time} to
 #'   optionally suppress warning messages.
@@ -52,19 +52,33 @@
 #' @export
 #'
 #' @examples
-#' starttime <- parseDatetime(2015080718, timezone = "America/Los_Angeles")
-#' datetimes <- parseDatetime(
-#'   c("20181014 12", "20181015 12", "20181016 12"),
+#' # All y[md-hms] formats are accepted
+#' parseDatetime(2018, timezone = "America/Los_Angeles")
+#' parseDatetime(201808, timezone = "America/Los_Angeles")
+#' parseDatetime(20180807, timezone = "America/Los_Angeles")
+#' parseDatetime(2018080718, timezone = "America/Los_Angeles")
+#' parseDatetime(201808071812, timezone = "America/Los_Angeles")
+#' parseDatetime(20180807181215, timezone = "America/Los_Angeles")
+#' parseDatetime("2018-08-07 18:12:15", timezone = "America/Los_Angeles")
+#'
+#' # Julian days are accepeted
+#' parseDatetime(2018219181215, timezone = "America/Los_Angeles",
+#'               isJulian = TRUE)
+#'
+#' # Vector dates are accepted and daylight savings is respected
+#' parseDatetime(
+#'   c("2018-10-24 12:00", "2018-10-31 12:00",
+#'     "2018-11-07 12:00", "2018-11-08 12:00"),
 #'   timezone = "America/New_York"
 #' )
 #'
-#' \dontrun{
 #' badInput <- c("20181013", NA, "20181015", "181016", "10172018")
 #'
-#' # This will return a vector with the date that were able to parse
+#' # Return a vector with \code{NA} for dates that could not be parsed
 #' parseDatetime(badInput, timezone = "UTC", expectAll = FALSE)
 #'
-#' # This will return an error, since some non-NA indices didn't parse
+#' \dontrun{
+#' # Fail if any dates cannot be parsed
 #' parseDatetime(badInput, timezone = "UTC", expectAll = TRUE)
 #' }
 #'
@@ -73,7 +87,7 @@ parseDatetime <- function(
   datetime = NULL,
   timezone = NULL,
   expectAll = FALSE,
-  julian = FALSE,
+  isJulian = FALSE,
   quiet = TRUE
 ) {
 
@@ -82,7 +96,7 @@ parseDatetime <- function(
   stopIfNull(datetime)
   stopIfNull(timezone)
   stopIfNull(expectAll)
-  stopIfNull(julian)
+  stopIfNull(isJulian)
   stopIfNull(quiet)
 
   if ( !is.character(timezone) || length(timezone) > 1 )
@@ -94,8 +108,8 @@ parseDatetime <- function(
   if ( !is.logical(expectAll) || length(expectAll) != 1 )
     stop("argument 'expectAll' must be a logical value of length one.")
 
-  if ( !is.logical(julian) || length(julian) != 1 )
-    stop("argument 'julian' must be a logical value of length one.")
+  if ( !is.logical(isJulian) || length(isJulian) != 1 )
+    stop("argument 'isJulian' must be a logical value of length one.")
 
   # Return early if already POSIXct -----------------------------------------
 
@@ -104,7 +118,7 @@ parseDatetime <- function(
 
   # Parse datetimes ---------------------------------------------------------
 
-  if ( julian ) {
+  if ( isJulian ) {
 
     # NOTE:  Julian date strings created by NASA satellite products often
     # NOTE:  include a digit for fractional seconds but no "." decimal marker.
@@ -120,7 +134,7 @@ parseDatetime <- function(
       paste0(wholePart, ".", fractionalPart) %>%
       stringr::str_replace("\\.$", "")
 
-    orders <- c("Yj", "YjH", "YjHM", "YjHMS")
+    orders <- c("Y", "Yj", "YjH", "YjHM", "YjHMS")
     parsedDatetime <- lubridate::parse_date_time(datetime,
                                                  orders,
                                                  tz = timezone,