+++ R/utils-flog.logger.R
@@ -1,0 +1,9 @@
+#' @title Manage loggers
+#'
+#' @description See \code{futile.logger::\link[futile.logger]{flog.logger}} for details.
+#'
+#' @name flog.logger
+#' @keywords internal
+#' @export
+#' @importFrom futile.logger flog.logger
+NULL